pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tago"
version = "0.1.0"
description = "AWS Tag Orchestrator"
requires-python = ">=3.11"
dependencies = [
  "boto3",
  "typer[all]",
  "jinja2",
  "pyyaml",
]

# aqui definimos o comando `tago`
[project.scripts]
tago = "cli.main:app"

# diz que os pacotes Python (cli, core, etc.) estão dentro da pasta src
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
